chore: unify user services
