--- src/modules/user/create-user/create-user.dto.ts
@@ -1,29 +1,0 @@
-import {
-  IsDateString,
-  IsEmail,
-  IsNotEmpty,
-  IsString,
-  IsStrongPassword,
-} from 'class-validator';
-
-export class CreateUserDto {
-  @IsNotEmpty()
-  @IsString()
-  name: string;
-
-  @IsNotEmpty()
-  @IsEmail()
-  email: string;
-
-  @IsString()
-  @IsStrongPassword()
-  password: string;
-
-  @IsNotEmpty()
-  @IsString()
-  passwordConfirmation: string;
-
-  @IsNotEmpty()
-  @IsDateString({ strict: true })
-  birthDate: string;
-}
