+++ src/modules/user/user.module.ts
@@ -1,15 +1,14 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { MailModule } from '../mail/mail.module';
-import { CreateUserController } from './create-user/create-user.controller';
-import { CreateUserService } from './create-user/create-user.service';
-import { GetUserService } from './get-user/get-user.service';
+import { UserController } from './user.controller';
 import { UserRepository } from './user.repository';
+import { UserService } from './user.service';
 
 @Module({
   imports: [PrismaModule, MailModule],
-  controllers: [CreateUserController],
-  providers: [CreateUserService, UserRepository, GetUserService],
-  exports: [UserRepository, GetUserService],
+  controllers: [UserController],
+  providers: [UserRepository, UserService],
+  exports: [UserRepository, UserService],
 })
 export class UserModule {}
